+++ src/empirical_standards/panel/__init__.py
@@ -2,5 +2,13 @@
 
 from empirical_standards.panel.fixed_effects import FixedEffectsResult, fit_fixed_effects
 from empirical_standards.panel.iv import PanelIV2SLSResult, fit_panel_iv_2sls
+from empirical_standards.panel.r_iv import RPanelIV2SLSResult, fit_panel_iv_2sls_r
 
-__all__ = ["FixedEffectsResult", "PanelIV2SLSResult", "fit_fixed_effects", "fit_panel_iv_2sls"]
+__all__ = [
+    "FixedEffectsResult",
+    "PanelIV2SLSResult",
+    "RPanelIV2SLSResult",
+    "fit_fixed_effects",
+    "fit_panel_iv_2sls",
+    "fit_panel_iv_2sls_r",
+]